made err in buff pointer to error
my thought behind this was you may need to handle multiple buffers in the same context.

and when one jumps to the error handling code it would all get very anoying so this would fix that also this does not apply to conn becouse that is the conntext object get it?

conn text context iam very funny i know